+++ src/list.c
@@ -65,7 +65,7 @@ xmlSecPtrListCreate(xmlSecPtrListId id) {
                     NULL,
                     XMLSEC_ERRORS_R_MALLOC_FAILED,
                     "sizeof(xmlSecPtrList)=%d",
-                    sizeof(xmlSecPtrList));
+                    (int)sizeof(xmlSecPtrList));
         return(NULL);
     }
 
@@ -479,7 +479,7 @@ xmlSecPtrListEnsureSize(xmlSecPtrListPtr list, xmlSecSize size) {
                     NULL,
                     XMLSEC_ERRORS_R_MALLOC_FAILED,
                     "sizeof(xmlSecPtr)*%d=%d",
-                    newSize, sizeof(xmlSecPtr) * newSize);
+                    newSize, (int)(sizeof(xmlSecPtr) * newSize));
         return(-1);
     }
